Further improvements to edit url in grids, now works on older Magento versions.
