+++ composer.json
@@ -18,6 +18,7 @@
         "magento/module-config": "^100.1 || ^101.0",
         "magento/module-cron": "^100.1",
         "magento/module-store": "^100.1 || ^101.0",
+        "magento/module-theme": "^100.1 || ^101.0",
         "magento/module-ui": "^100.1 || ^101.0",
         "symfony/console": "^2.5 || ^3.0 || ^4.0 || ^5.0"
     },
