+++ view/adminhtml/ui_component/baldwin_urldataintegritychecker_grid_catalog_product_urlkey.xml
@@ -69,10 +69,13 @@
                 </item>
             </argument>
         </column>
-        <actionsColumn name="actions" class="Baldwin\UrlDataIntegrityChecker\Ui\Component\Listing\Columns\ProductActions">
-            <settings>
-                <indexField>productId</indexField>
-            </settings>
+        <actionsColumn name="actions" class="Baldwin\UrlDataIntegrityChecker\Ui\Component\Listing\Column\EditAction">
+            <argument name="data" xsi:type="array">
+                <item name="config" xsi:type="array">
+                    <item name="indexField" xsi:type="string">productId</item>
+                    <item name="editUrlPath" xsi:type="string">catalog/product/edit</item>
+                </item>
+            </argument>
         </actionsColumn>
     </columns>
 </listing>
